+++ packages/sdk/src/index.ts
@@ -1,6 +1,7 @@
-export * from "@nocturne-xyz/base-utils";
-export * from "@nocturne-xyz/primitives";
-
+export * from "./utils";
+export * from "./primitives";
+export * from "./proof";
+export * from "./crypto";
 export * from "./db";
 export * from "./indexing";
 export * from "./merkleProver";