+++ packages/sdk/src/primitives/note.ts
@@ -1,8 +1,8 @@
-import { StealthAddressTrait, StealthAddress } from "./crypto";
+import { StealthAddressTrait, StealthAddress } from "../crypto";
 import { Asset, AssetTrait, EncodedAsset } from "./asset";
 import { poseidonBN } from "@nocturne-xyz/circuit-utils";
 import { sha256 } from "js-sha256";
-import { bigintToBEPadded } from "@nocturne-xyz/base-utils";
+import { bigintToBEPadded } from "../utils";
 
 export interface Note {
   owner: StealthAddress;