+++ packages/sdk/src/utils/misc.ts
@@ -1,4 +1,4 @@
-import { Note } from "@nocturne-xyz/primitives";
+import { Note } from "../primitives";
 
 export function sortNotesByValue<T extends Note>(notes: T[]): T[] {
   return notes.sort((a, b) => {